+++ crates/zng-wgt-layer/src/popup.rs
@@ -228,10 +228,21 @@ impl POPUP {
         )
         .boxed();
 
-        if let ContextCapture::CaptureBlend { filter, over } = context_capture {
-            if filter != CaptureFilter::None {
-                popup = with_context_blend(LocalContext::capture_filtered(filter), over, popup).boxed();
-            }
+        let (filter, over) = match context_capture {
+            ContextCapture::NoCapture => {
+                let filter = CaptureFilter::Include({
+                    let mut set = ContextValueSet::new();
+                    set.insert(&CLOSE_ON_FOCUS_LEAVE_VAR);
+                    set.insert(&ANCHOR_MODE_VAR);
+                    set.insert(&CONTEXT_CAPTURE_VAR);
+                    set
+                });
+                (filter, false)
+            },
+            ContextCapture::CaptureBlend { filter, over } => (filter, over),
+        };
+        if filter != CaptureFilter::None {
+            popup = with_context_blend(LocalContext::capture_filtered(filter), over, popup).boxed();
         }
         LAYERS.insert_anchored(LayerIndex::TOP_MOST, anchor_id, anchor_mode, popup);
 
@@ -335,8 +346,11 @@ impl DefaultStyle {
 /// [`with_context_blend`]: zng_wgt::prelude::with_context_blend
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub enum ContextCapture {
-    /// No context capture or blending, the popup will have
-    /// the context it is inited in, like any other widget.
+    /// No context capture except the popup configuration context.
+    /// 
+    /// The popup will only have the window context as it is open as a layer on the window root.
+    /// 
+    /// Note to filter out even the popup config use [`CaptureFilter::None`] instead.
     NoCapture,
     /// Build/instantiation context is captured and blended with the node context during all [`UiNodeOp`].
     ///